Split implementation out of BehaviorTester
