+++ src/test/java/org/inferred/freebuilder/processor/BuildablePropertyFactoryTest.java
@@ -206,7 +206,7 @@ public class BuildablePropertyFactoryTest {
       .build();
 
   @Rule public final ExpectedException thrown = ExpectedException.none();
-  private final BehaviorTester behaviorTester = new BehaviorTester();
+  private final BehaviorTester behaviorTester = BehaviorTester.create();
 
   @Test
   public void testBuild_noDefaults() {
