+++ src/test/java/org/inferred/freebuilder/processor/DefaultMapperMethodTest.java
@@ -57,7 +57,7 @@ public class DefaultMapperMethodTest {
       .build();
 
   @Rule public final ExpectedException thrown = ExpectedException.none();
-  private final BehaviorTester behaviorTester = new BehaviorTester();
+  private final BehaviorTester behaviorTester = BehaviorTester.create();
 
   @Test
   public void mapReplacesValueToBeReturnedFromGetterForRequiredProperty() {
