+++ src/test/java/org/inferred/freebuilder/processor/JavaUtilOptionalPropertyTest.java
@@ -82,7 +82,7 @@ public class JavaUtilOptionalPropertyTest {
       .build();
 
   @Rule public final ExpectedException thrown = ExpectedException.none();
-  private final BehaviorTester behaviorTester = new BehaviorTester();
+  private final BehaviorTester behaviorTester = BehaviorTester.create();
 
   @Test
   public void testConstructor_defaultEmpty() {
