+++ src/test/java/org/inferred/freebuilder/processor/ListMultimapPropertyFactoryTest.java
@@ -99,7 +99,7 @@ public class ListMultimapPropertyFactoryTest {
       .build();
 
   @Rule public final ExpectedException thrown = ExpectedException.none();
-  private final BehaviorTester behaviorTester = new BehaviorTester();
+  private final BehaviorTester behaviorTester = BehaviorTester.create();
 
   @Test
   public void testDefaultEmpty() {
