+++ src/test/java/org/inferred/freebuilder/processor/ListMutateMethodTest.java
@@ -76,7 +76,7 @@ public class ListMutateMethodTest {
       .build();
 
   @Rule public final ExpectedException thrown = ExpectedException.none();
-  private final BehaviorTester behaviorTester = new BehaviorTester();
+  private final BehaviorTester behaviorTester = BehaviorTester.create();
 
   @Test
   public void mutateAndAddModifiesUnderlyingPropertyWhenUnchecked() {
