+++ src/test/java/org/inferred/freebuilder/processor/ListPropertyFactoryTest.java
@@ -54,7 +54,7 @@ public class ListPropertyFactoryTest {
       .build();
 
   @Rule public final ExpectedException thrown = ExpectedException.none();
-  private final BehaviorTester behaviorTester = new BehaviorTester();
+  private final BehaviorTester behaviorTester = BehaviorTester.create();
 
   @Test
   public void testDefaultEmpty() {
