+++ src/test/java/org/inferred/freebuilder/processor/MapMutateMethodTest.java
@@ -64,7 +64,7 @@ public class MapMutateMethodTest {
       .build();
 
   @Rule public final ExpectedException thrown = ExpectedException.none();
-  private final BehaviorTester behaviorTester = new BehaviorTester();
+  private final BehaviorTester behaviorTester = BehaviorTester.create();
 
   @Test
   public void putModifiesUnderlyingPropertyWhenUnchecked() {
