+++ src/test/java/org/inferred/freebuilder/processor/MapPropertyFactoryTest.java
@@ -81,7 +81,7 @@ public class MapPropertyFactoryTest {
       .build();
 
   @Rule public final ExpectedException thrown = ExpectedException.none();
-  private final BehaviorTester behaviorTester = new BehaviorTester();
+  private final BehaviorTester behaviorTester = BehaviorTester.create();
 
   @Test
   public void testDefaultEmpty() {
