+++ src/test/java/org/inferred/freebuilder/processor/MultisetMutateMethodTest.java
@@ -76,7 +76,7 @@ public class MultisetMutateMethodTest {
       .build();
 
   @Rule public final ExpectedException thrown = ExpectedException.none();
-  private final BehaviorTester behaviorTester = new BehaviorTester();
+  private final BehaviorTester behaviorTester = BehaviorTester.create();
 
   @Test
   public void mutateAndAddModifiesUnderlyingProperty_whenUnchecked() {
