+++ src/test/java/org/inferred/freebuilder/processor/NullableMapperMethodTest.java
@@ -44,7 +44,7 @@ public class NullableMapperMethodTest {
       .build();
 
   @Rule public final ExpectedException thrown = ExpectedException.none();
-  private final BehaviorTester behaviorTester = new BehaviorTester();
+  private final BehaviorTester behaviorTester = BehaviorTester.create();
 
   @Test
   public void replacesValueToBeReturnedFromGetter() {
