+++ src/test/java/org/inferred/freebuilder/processor/NullablePropertyTest.java
@@ -74,7 +74,7 @@ public class NullablePropertyTest {
       .build();
 
   @Rule public final ExpectedException thrown = ExpectedException.none();
-  private final BehaviorTester behaviorTester = new BehaviorTester();
+  private final BehaviorTester behaviorTester = BehaviorTester.create();
 
   @Test
   public void testConstructor_defaultAbsent() {
