+++ src/test/java/org/inferred/freebuilder/processor/ProcessorTest.java
@@ -99,7 +99,7 @@ public class ProcessorTest {
       .build();
 
   @Rule public final ExpectedException thrown = ExpectedException.none();
-  private final BehaviorTester behaviorTester = new BehaviorTester();
+  private final BehaviorTester behaviorTester = BehaviorTester.create();
 
   @Test
   public void testAbstractClass() {
