+++ src/test/java/org/inferred/freebuilder/processor/RequiredPropertiesTest.java
@@ -61,7 +61,7 @@ public class RequiredPropertiesTest {
       .build();
 
   @Rule public final ExpectedException thrown = ExpectedException.none();
-  private final BehaviorTester behaviorTester = new BehaviorTester();
+  private final BehaviorTester behaviorTester = BehaviorTester.create();
 
   @Test
   public void testCantBuildWithAnUnsetProperty() {
