+++ src/test/java/org/inferred/freebuilder/processor/SetMultimapMutateMethodTest.java
@@ -81,7 +81,7 @@ public class SetMultimapMutateMethodTest {
       .build();
 
   @Rule public final ExpectedException thrown = ExpectedException.none();
-  private final BehaviorTester behaviorTester = new BehaviorTester();
+  private final BehaviorTester behaviorTester = BehaviorTester.create();
 
   @Test
   public void mutateAndPutModifiesUnderlyingProperty_whenUnchecked() {
