+++ src/test/java/org/inferred/freebuilder/processor/SetMultimapPropertyFactoryTest.java
@@ -98,7 +98,7 @@ public class SetMultimapPropertyFactoryTest {
       .build();
 
   @Rule public final ExpectedException thrown = ExpectedException.none();
-  private final BehaviorTester behaviorTester = new BehaviorTester();
+  private final BehaviorTester behaviorTester = BehaviorTester.create();
 
   @Test
   public void testDefaultEmpty() {
