+++ src/test/java/org/inferred/freebuilder/processor/SetMutateMethodTest.java
@@ -63,7 +63,7 @@ public class SetMutateMethodTest {
       .build();
 
   @Rule public final ExpectedException thrown = ExpectedException.none();
-  private final BehaviorTester behaviorTester = new BehaviorTester();
+  private final BehaviorTester behaviorTester = BehaviorTester.create();
 
   @Test
   public void mutateAndAddModifiesUnderlyingProperty() {
