+++ src/test/java/org/inferred/freebuilder/processor/SetPropertyFactoryTest.java
@@ -69,7 +69,7 @@ public class SetPropertyFactoryTest {
       .build();
 
   @Rule public final ExpectedException thrown = ExpectedException.none();
-  private final BehaviorTester behaviorTester = new BehaviorTester();
+  private final BehaviorTester behaviorTester = BehaviorTester.create();
 
   @Test
   public void testDefaultEmpty() {
