+++ src/test/java/org/inferred/freebuilder/processor/util/testing/TestBuilder.java
@@ -26,7 +26,7 @@ import javax.tools.JavaFileObject;
 import javax.tools.SimpleJavaFileObject;
 
 /**
- * Simple builder API for a test method, suitable for use in {@link BehaviorTester}. See the
+ * Simple builder API for a test method, suitable for use in {@link SingleBehaviorTester}. See the
  * JavaDoc on that class for an example.
  *
  * <p>Automatically imports {@link Assert}.* and {@link Truth#assertThat}.
